POW 2o3 - v1.0.2
Change all integer inputs to floats.

+++ Extension Objects/Full Extensions/_Source Codes/DarkEdif/POW 2o3/Extension.hpp
@@ -79,9 +79,9 @@ public:
 
 	// Expressions
 
-	int SquareParam(int IntegerInput);
-	int CubeParam(int IntegerInput);
-	int PowFunc(int MainInput, int PowerInput);
+	float SquareParam(float IntegerInput);
+	float CubeParam(float IntegerInput);
+	float PowFunc(float MainInput, float PowerInput);
 
 	// Runs every tick of Fusion's runtime, can be toggled off and back on
 	REFLAG Handle();
